requirements.txt
asgiref==3.8.1
Django==3.2.19
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==22.0.0
packaging==24.0
pymongo==3.12.3
python-dotenv==1.0.1
sqlparse==0.2.4
typing_extensions==4.11.0